backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
mistralai==1.2.5
httpx==0.27.0
python-dotenv==1.0.1
